simple-animations: Fix compiler warnings

+++ src/simple-animations/animationsim.c
@@ -349,7 +349,7 @@ animGetObjectOptions (CompPlugin *plugin,
     };
 
     RETURN_DISPATCH (object, dispTab, ARRAY_SIZE (dispTab),
-		     (void *) (*count = 0), (plugin, object, count));
+		     (void *) (long int)(*count = 0), (plugin, object, count));
 }
 
 static CompBool
